+++ CHANGELOG.md
@@ -4,6 +4,11 @@
  * Loading a lottie with missing properties will retain the default value
  * Fixed parsing for lottie files with unknown effects
  * Parser for after effects .aep and .aepx project files
+ * Renamed `objects.easing.Jump` to `objects.easing.Hold` (the old name will still work)
+ * Added `objects.easing.Bezier`
+ * `add_keyframe()` not returns the keyframe being added
+ * `add_keyframe()` will no longer set the old-fashioned `end` value unless you pass `end=True`
+ * Renamed keyframe `.start` to `.value` (the old name will still work)
 
 ## 0.7.0 - 2022-09-11
  * Improved color management